Stop waiting if workflow has been superseded
This change is helpful for longer wait intervals, for example 30 minutes, by checking if the workflow has been superseded at least every minute (and at most every second, for smaller intervals), and stop waiting if it has.

I'm using this action as part of the deploy workflow to wait for 30 minutes between different deployment targets.

Thanks for the good work on this action!